Support for NAT and port forwarding
The Rex Protocol assumes a local network, one-ip-per-device, and all devices listening to standard UDP ports.

But someone is trying to use a series of devices behind a NAT, with different exposed ports forwarding to each device.

To work around that, a new override of `RexClient.RequestConnection()` was created, receiving:
- The IP/Port of the Rex's UDP socket for connection requests -> This should forward to the Rex's IP, port 1025.
- The IP/Port of the Server's TCP socket -> This should be the IP/Port acessible by the Rex that maps to the local IP and port,